chore: move setField to style to pass codacy scan
When move setField to utils.go Codacy will rise the security issue
unsafe-reflect-by-name. But actually all the fields and mapping options
are define as constant in this library, user can't input their custom map to map
value. So it is not a security issue. But when I write setField in
style.go, Codacy will think this is not a security (maybe because there
include some functions to use the setField so Codacy think it is safe)
issue, therefore I move setField back to style.go temporarily.